# Implement two-phase diagnostic publishing for immediate syntax error feedback