… CLAUDE.md
… docs/SERVER_ARCHITECTURE.md
… include/slangd/semantic/diagnostic_index.hpp
… include/slangd/services/session_manager.hpp
… src/slangd/core/slangd_lsp_server.cpp
… src/slangd/semantic/diagnostic_converter.cpp
… src/slangd/services/language_service.cpp
… src/slangd/services/overlay_session.cpp
… src/slangd/services/session_manager.cpp
… test/slangd/common/simple_fixture.cpp
… test/slangd/common/simple_fixture.hpp
… test/slangd/semantic/BUILD.bazel
… test/slangd/semantic/diagnostic_converter_test.cpp