## Summary

Introduces two-phase diagnostic publishing to provide immediate syntax error feedback (~7ms) while semantic analysis runs in background (~500ms). Creates SessionManager for centralized session lifecycle management, eliminating race conditions and duplicate compilation overhead.

**User experience:**
- **Before**: Save → wait 2-3s → see diagnostics
- **After**: Save → syntax errors appear instantly (~7ms) → full diagnostics after elaboration (~500ms)

## Key Performance Improvements

**Immediate Parse Diagnostics**: 7ms (was 848ms)
- Single-file compilation without catalog dependencies
- Users see syntax errors instantly on save

**Optimized Full Diagnostics**: 126ms compilation + background indexing
- BufferID-based filtering (0ms for 7950 diagnostics, was string-based path comparison)
- Two-phase channels allow diagnostics without waiting for indexing

**Eliminated Race Conditions**: 
- Concurrent requests wait for first compilation (20ms) instead of duplicating work (480ms)
- LSP document version tracking prevents stale session updates
- Proper async coordination ensures session creation completes before feature requests

## SessionManager Architecture

**Centralized Session Lifecycle** (NEW):
- Event-driven session creation with async channels
- LRU cache keyed by URI (stable during typing)
- Two-phase channels signal `compilation_ready` vs `session_ready` states
- Prevents duplicate compilation when multiple requests arrive concurrently

**Ownership Hierarchy**:
```
SlangdLspServer (LSP protocol layer)
  └── LanguageService (public API, business logic)
        └── SessionManager (private resource management)
              └── OverlaySession (data: compilation + index)
```

**Protocol-Level Event Handlers**:
- `OnDocumentOpened/Saved` → triggers `SessionManager.UpdateSession()` (private)
- `OnDocumentClosed` → triggers `SessionManager.RemoveSession()` (private)
- `OnDocumentsChanged` → triggers `SessionManager.InvalidateSessions()` (private)

**Why SessionManager?**
- Centralizes all session creation, caching, and invalidation logic
- Features are read-only consumers - never create sessions directly
- Async channels coordinate concurrent requests at different lifecycle phases
- Clean separation: LanguageService owns SessionManager, SessionManager owns OverlaySessions

## Diagnostic Publishing Strategy

**Two-Phase Approach**:
1. **Parse diagnostics** publish immediately (syntax errors, no elaboration required)
2. **Full diagnostics** publish after elaboration (semantic analysis)

**Both phases reuse the same cached OverlaySession** - first phase may create session, second phase reuses it. This follows Slang's lazy evaluation pattern.

**Key architectural decision**: Diagnostic extraction separated from session creation to enable lazy elaboration. Sessions no longer force compilation during creation.

## Architecture Refactoring

### DiagnosticConverter (was DiagnosticIndex)
- Changed from stateful index to stateless utility class
- `ExtractParseDiagnostics()` - syntax errors only, no elaboration
- `ExtractAllDiagnostics()` - full semantic analysis with elaboration
- Diagnostic extraction now on-demand, not during session creation

### LanguageService API Clarity
Renamed methods to clarify they're protocol-level events, not session management:
- `UpdateSession` → `OnDocumentOpened/OnDocumentSaved`
- `RemoveSession` → `OnDocumentClosed`
- `InvalidateSessions` → `OnDocumentsChanged`

Clear separation: SlangdLspServer calls protocol events, LanguageService manages SessionManager internally.

### OverlaySession Improvements
- Added `BuildCompilation()` factory method for flexible session creation
- Stores `BufferID` for O(1) diagnostic filtering (eliminates string path comparison)
- Added `GetCompilationPtr()` for async lifetime management in `CompilationState`
- Removed old cache key infrastructure (`OverlayCacheKey`)

### Complete Migration to SessionManager
- Removed old `overlay_cache_` and `GetOrCreateOverlay` logic
- All features now use `SessionManager.GetSession()` for read-only access
- `ComputeDiagnostics` uses `SessionManager.GetSessionForDiagnostics()` for fast compilation state
- Simplified function signatures by removing content parameters

## Bug Fixes

### Race Condition in OnDidSaveTextDocument
- **Issue**: Detached `co_spawn` caused diagnostics to run before session was ready
- **Fix**: Now awaits `UpdateSession` completion before publishing diagnostics
- **Impact**: Ensures session is fully created before features can access it

### Race Condition in SessionManager
- **Issue**: Concurrent updates could clobber each other's sessions
- **Fix**: LSP document version tracking - only erase pending sessions matching current version
- **Impact**: Stale updates from previous document versions don't overwrite newer sessions

### File Watcher Bug
- **Issue**: File watcher invalidated sessions for open files (LSP already manages those)
- **Fix**: External changes only affect closed files
- **Impact**: Stopped unnecessary session rebuilds for files being edited

### OnDidOpen Race Condition
- **Issue**: Session creation didn't complete before handler returned
- **Fix**: Await `UpdateSession` before returning from `OnDidOpen`
- **Impact**: Features can safely access sessions immediately after open event

## Documentation

Extensive updates to `SERVER_ARCHITECTURE.md`:
- Two-phase diagnostic publishing strategy with timing breakdowns
- SessionManager ownership hierarchy and event-driven model
- Async channel architecture for diagnostic optimization (compilation_ready vs session_ready)
- Extension guidelines for adding new LSP features
- Clear separation between protocol events and session management

Added to `CLAUDE.md`:
- SessionManager note in architecture section
- Key principle: features are read-only consumers, never create sessions directly

## Testing

- [x] All 17 tests passing
- [x] Manual E2E testing verified immediate syntax error feedback
- [x] Performance logging confirms timing improvements (parse: 7ms, full: 126ms compilation)
- [x] Race condition fixes verified with concurrent request testing
- [x] Added modern C++23 test helpers using `std::ranges` for cleaner assertions